+++ src/main/java/com/example/surveyapp/domain/admin/service/AdminService.java
@@ -70,7 +70,7 @@ public class AdminService {
                 () -> new CustomException(ErrorCode.NOT_FOUND_USER)
         );
 
-        if (blackListRepository.findByUserId(user).isPresent()) {
+        if (blackListRepository.findByUserId(userId).isPresent()) {
             throw new CustomException(ErrorCode.IS_BLACKLIST);
         }
 
@@ -87,7 +87,7 @@ public class AdminService {
                 () -> new CustomException(ErrorCode.NOT_FOUND_USER)
         );
 
-        BlackList blackList = blackListRepository.findByUserId(user).orElseThrow(
+        BlackList blackList = blackListRepository.findByUserId(userId).orElseThrow(
                 () -> new CustomException(ErrorCode.IS_NOT_BLACKLIST)
         );
