--- src/main/java/com/example/surveyapp/domain/order/facade/UserFacade.java
@@ -1,7 +1,0 @@
-package com.example.surveyapp.domain.order.facade;
-
-import com.example.surveyapp.domain.user.domain.model.User;
-
-public interface UserFacade {
-    Long findUser(Long id);
-}
